+++ backend/pyproject.toml
@@ -25,6 +25,9 @@ dependencies = [
     "psycopg[binary]>=3.2.9",
     "passlib[bcrypt]>=1.7.4",
     "jinja2>=3.1.6",
+    "langgraph>=0.2.0",
+    "langchain>=0.3.0",
+    "langchain-openai>=0.2.0",
 ]
 
 [dependency-groups]
