--- backend/src/app/services/__init__.py
@@ -1,1 +1,0 @@
-# Services package
